requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
httpx
playwright
pytest
tzdata


